CoBienICAM-main/frontend-mueble/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
coloredlogs==15.0.1
datasets==3.5.0
dill==0.3.8
dnspython==2.7.0
docutils==0.21.2
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
joblib==1.4.2
Kivy==2.3.1
Kivy-Garden==0.1.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.4.2
multiprocess==0.70.16
networkx==3.4.2
numpy==1.24.3
onnxruntime==1.21.0
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.1.0
playsound==1.2.2
polib==1.2.0
propcache==0.3.1
protobuf==6.30.2
proxy_tools==0.1.0
pyarrow==19.0.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pygame
PyGetWindow==0.0.9
Pygments==2.19.1
pymongo==4.12.0
PyMsgBox==1.0.9
# pyobjc==11.0
# pyobjc-core==11.0
# pyobjc-framework-Accessibility==11.0
# pyobjc-framework-Accounts==11.0
# pyobjc-framework-AddressBook==11.0
# pyobjc-framework-AdServices==11.0
# pyobjc-framework-AdSupport==11.0
# pyobjc-framework-AppleScriptKit==11.0
# pyobjc-framework-AppleScriptObjC==11.0
# pyobjc-framework-ApplicationServices==11.0
# pyobjc-framework-AppTrackingTransparency==11.0
# pyobjc-framework-AudioVideoBridging==11.0
# pyobjc-framework-AuthenticationServices==11.0
# pyobjc-framework-AutomaticAssessmentConfiguration==11.0
# pyobjc-framework-Automator==11.0
# pyobjc-framework-AVFoundation==11.0
# pyobjc-framework-AVKit==11.0
# pyobjc-framework-AVRouting==11.0
# pyobjc-framework-BackgroundAssets==11.0
# pyobjc-framework-BusinessChat==11.0
# pyobjc-framework-CalendarStore==11.0
# pyobjc-framework-CallKit==11.0
# pyobjc-framework-Carbon==11.0
# pyobjc-framework-CFNetwork==11.0
# pyobjc-framework-ClassKit==11.0
# pyobjc-framework-CloudKit==11.0
# pyobjc-framework-Cocoa==11.0
# pyobjc-framework-Collaboration==11.0
# pyobjc-framework-ColorSync==11.0
# pyobjc-framework-Contacts==11.0
# pyobjc-framework-ContactsUI==11.0
# pyobjc-framework-CoreAudio==11.0
# pyobjc-framework-CoreAudioKit==11.0
# pyobjc-framework-CoreBluetooth==11.0
# pyobjc-framework-CoreData==11.0
# pyobjc-framework-CoreHaptics==11.0
# pyobjc-framework-CoreLocation==11.0
# pyobjc-framework-CoreMedia==11.0
# pyobjc-framework-CoreMediaIO==11.0
# pyobjc-framework-CoreMIDI==11.0
# pyobjc-framework-CoreML==11.0
# pyobjc-framework-CoreMotion==11.0
# pyobjc-framework-CoreServices==11.0
# pyobjc-framework-CoreSpotlight==11.0
# pyobjc-framework-CoreText==11.0
# pyobjc-framework-CoreWLAN==11.0
# pyobjc-framework-CryptoTokenKit==11.0
# pyobjc-framework-DataDetection==11.0
# pyobjc-framework-DeviceCheck==11.0
# pyobjc-framework-DictionaryServices==11.0
# pyobjc-framework-DiscRecording==11.0
# pyobjc-framework-DiscRecordingUI==11.0
# pyobjc-framework-DiskArbitration==11.0
# pyobjc-framework-DVDPlayback==11.0
# pyobjc-framework-EventKit==11.0
# pyobjc-framework-ExceptionHandling==11.0
# pyobjc-framework-ExecutionPolicy==11.0
# pyobjc-framework-ExtensionKit==11.0
# pyobjc-framework-ExternalAccessory==11.0
# pyobjc-framework-FileProvider==11.0
# pyobjc-framework-FileProviderUI==11.0
# pyobjc-framework-FinderSync==11.0
# pyobjc-framework-FSEvents==11.0
# pyobjc-framework-GameCenter==11.0
# pyobjc-framework-GameController==11.0
# pyobjc-framework-GameKit==11.0
# pyobjc-framework-GameplayKit==11.0
# pyobjc-framework-HealthKit==11.0
# pyobjc-framework-ImageCaptureCore==11.0
# pyobjc-framework-InputMethodKit==11.0
# pyobjc-framework-InstallerPlugins==11.0
# pyobjc-framework-InstantMessage==11.0
# pyobjc-framework-Intents==11.0
# pyobjc-framework-IntentsUI==11.0
# pyobjc-framework-IOBluetooth==11.0
# pyobjc-framework-IOBluetoothUI==11.0
# pyobjc-framework-IOSurface==11.0
# pyobjc-framework-iTunesLibrary==11.0
# pyobjc-framework-KernelManagement==11.0
# pyobjc-framework-LatentSemanticMapping==11.0
# pyobjc-framework-LaunchServices==11.0
# pyobjc-framework-libdispatch==11.0
# pyobjc-framework-libxpc==11.0
# pyobjc-framework-LinkPresentation==11.0
# pyobjc-framework-LocalAuthentication==11.0
# pyobjc-framework-LocalAuthenticationEmbeddedUI==11.0
# pyobjc-framework-MailKit==11.0
# pyobjc-framework-MapKit==11.0
# pyobjc-framework-MediaAccessibility==11.0
# pyobjc-framework-MediaLibrary==11.0
# pyobjc-framework-MediaPlayer==11.0
# pyobjc-framework-MediaToolbox==11.0
# pyobjc-framework-Metal==11.0
# pyobjc-framework-MetalFX==11.0
# pyobjc-framework-MetalKit==11.0
# pyobjc-framework-MetalPerformanceShaders==11.0
# pyobjc-framework-MetalPerformanceShadersGraph==11.0
# pyobjc-framework-MetricKit==11.0
# pyobjc-framework-MLCompute==11.0
# pyobjc-framework-ModelIO==11.0
# pyobjc-framework-MultipeerConnectivity==11.0
# pyobjc-framework-NaturalLanguage==11.0
# pyobjc-framework-NetFS==11.0
# pyobjc-framework-Network==11.0
# pyobjc-framework-NetworkExtension==11.0
# pyobjc-framework-NotificationCenter==11.0
# pyobjc-framework-OpenDirectory==11.0
# pyobjc-framework-OSAKit==11.0
# pyobjc-framework-OSLog==11.0
# pyobjc-framework-PassKit==11.0
# pyobjc-framework-PencilKit==11.0
# pyobjc-framework-PHASE==11.0
# pyobjc-framework-Photos==11.0
# pyobjc-framework-PhotosUI==11.0
# pyobjc-framework-PreferencePanes==11.0
# pyobjc-framework-PushKit==11.0
# pyobjc-framework-Quartz==11.0
# pyobjc-framework-QuickLookThumbnailing==11.0
# pyobjc-framework-ReplayKit==11.0
# pyobjc-framework-SafariServices==11.0
# pyobjc-framework-SafetyKit==11.0
# pyobjc-framework-SceneKit==11.0
# pyobjc-framework-ScreenCaptureKit==11.0
# pyobjc-framework-ScreenSaver==11.0
# pyobjc-framework-ScreenTime==11.0
# pyobjc-framework-ScriptingBridge==11.0
# pyobjc-framework-SearchKit==11.0
# pyobjc-framework-Security==11.0
# pyobjc-framework-SecurityFoundation==11.0
# pyobjc-framework-SecurityInterface==11.0
# pyobjc-framework-ServiceManagement==11.0
# pyobjc-framework-SharedWithYou==11.0
# pyobjc-framework-SharedWithYouCore==11.0
# pyobjc-framework-ShazamKit==11.0
# pyobjc-framework-Social==11.0
# pyobjc-framework-SoundAnalysis==11.0
# pyobjc-framework-Speech==11.0
# pyobjc-framework-SpriteKit==11.0
# pyobjc-framework-StoreKit==11.0
# pyobjc-framework-SyncServices==11.0
# pyobjc-framework-SystemConfiguration==11.0
# pyobjc-framework-SystemExtensions==11.0
# pyobjc-framework-ThreadNetwork==11.0
# pyobjc-framework-UniformTypeIdentifiers==11.0
# pyobjc-framework-UserNotifications==11.0
# pyobjc-framework-UserNotificationsUI==11.0
# pyobjc-framework-VideoSubscriberAccount==11.0
# pyobjc-framework-VideoToolbox==11.0
# pyobjc-framework-Virtualization==11.0
# pyobjc-framework-Vision==11.0
# pyobjc-framework-WebKit==11.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.17
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-vlc==3.0.21203
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.2
pywebview==5.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rubicon-objc==0.5.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.7
srt==3.5.3
sympy==1.13.3
threadpoolctl==3.6.0
timezonefinder
tokenizers==0.21.1
torch==2.1.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
vosk==0.3.44
websockets==15.0.1
Wikipedia-API==0.8.1
xxhash==3.5.0
yarl==1.19.0